fix: change bus_seat_price float to double

+++ src/main/java/org/transportreservation/model/BusSeat.java
@@ -11,7 +11,7 @@ import lombok.*;
 
 public class BusSeat {
     private int id_bus_seat;
-    private float bus_seat_price;
+    private double bus_seat_price;
     private boolean is_taken;
     private int id_bus;
 }
